+++ src/shell.ts
@@ -1,0 +1,36 @@
+/*
+ * Copyright 2022 steadybit GmbH. All rights reserved.
+ */
+
+import { exec as execWithoutPromises } from 'child_process';
+import { promisify } from 'util';
+
+const exec = promisify(execWithoutPromises);
+
+export interface ShellCommandResult {
+  stdout: string;
+  stderr: string;
+}
+
+export interface ShellCommandOptions {
+  trimOutput?: boolean;
+  cwd?: string;
+  timeout?: number;
+}
+
+export async function executeShellCommand(
+  command: string,
+  { trimOutput = true, cwd, timeout = 60_000 }: ShellCommandOptions = {}
+): Promise<ShellCommandResult> {
+  const result = await exec(command, {
+    cwd,
+    timeout,
+  });
+
+  if (trimOutput) {
+    result.stdout = result.stdout.trim();
+    result.stderr = result.stderr.trim();
+  }
+
+  return result;
+}